tests/wpt/tests/navigation-api/navigate-event/navigate-meta-refresh.html

<!doctype html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="resources/meta-refresh.html"></iframe>
</head>
<script>
async_test(t => {
  i.onload = t.step_func(() => {
    i.contentWindow.navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "reload");
      assert_true(e.cancelable);
      assert_true(e.canIntercept);
      assert_false(e.userInitiated);
      assert_false(e.hashChange);
      assert_equals(e.downloadRequest, null);
      assert_equals(e.destination.url, i.contentWindow.location.href);
      assert_false(e.destination.sameDocument);
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      assert_equals(e.formData, null);
      // NavigateEvent sourceElement is still in development, so test whether it is available.
      if ("sourceElement" in e) assert_equals(e.sourceElement, null);
      e.preventDefault();
    });
  });
}, "<meta> refresh fires navigate event");
</script>
